data_scrapping/requirements.txt
numpy==1.23.4
requests==2.22.0
pandas==1.5.1
folium==0.13.0
geopy==2.2.0
Levenshtein==0.20.8
beautifulsoup4==4.11.1
Pillow==9.3.0
plotly==5.11.0
lxml==4.9.1
openpyxl==3.0.10